- refactor remaining performance prediction code, address issues with hybrid parallelization - re-implement blocking factor calculation

+++ Compiler/src/exastencils/config/Platform.scala
@@ -176,7 +176,7 @@ object Platform {
   // in B
   var hw_cacheLineSize : Int = 512
   var hw_numCaches : Int = 1
-  var hw_cacheSize : Int = 32 * 1000 * 1000
+  var hw_cacheSize : Double = 32 * 1000 * 1000
   var hw_usableCache : Double = 1.0 // as multiplicative factor
   var hw_numCacheSharingThreads : Int = 4
   var hw_gpu_name : String = "NVidia Quadro 4000"
